--- api/src/modules/plots/plot/strategies/histogram.py
@@ -1,13 +1,0 @@
-from pandas import DataFrame
-from matplotlib import pyplot as plt
-from src.modules.plots.plot.strategies.strategy import PlotStrategy
-from src.modules.plots.data import descriptions
-
-class Histogram(PlotStrategy):
-    def __init__(self, column: str):
-        super().__init__(path=f"{descriptions[column].replace(' ', '-')}/histogram.png")
-        self.column = column
-    
-    def plot(self, dataframe: DataFrame):
-        plt.hist(dataframe[self.column], bins=30, color="blue", edgecolor="black", alpha=0.5)
-        self.set_labels(descriptions[self.column], descriptions["FREQ"])
